Remove deprecated types and dead code
